requirements.txt
asgiref==3.9.1
Django==5.2.4
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
celery==5.5.3
redis==5.2.1
django-celery-beat==2.6.0
flower==2.0.1
playwright==1.48.0
kombu==5.5.4
billiard==4.2.1
click==8.2.1
click-didyoumean==0.3.1
click-repl==0.3.0
click-plugins==1.1.1.2
python-dateutil==2.9.0.post0
amqp==5.3.1
tzdata==2025.2
packaging==25.0
prompt-toolkit==3.0.51
wcwidth==0.2.13
six==1.17.0
tornado==6.5.1
prometheus-client==0.22.1
humanize==4.12.3
pytz==2025.2
djangorestframework==3.16.0
